refactor: move files to autoimported shared/utils
