+++ shared/utils/calc-linear-progression.ts
